refactor: split dashboard metric data helpers
